requirements.txt
# Web Framework
fastapi<0.100.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# OpenAI
openai>=1.12.0,<1.30.0

# Email Processing
imapclient>=2.3.1
email-validator>=2.1.0
mail-parser>=3.15.0

# File Processing
python-docx>=1.0.1
openpyxl>=3.1.2
python-pptx>=0.6.22
PyPDF2>=3.0.1
pdf2image>=1.17.0
Pillow>=10.1.0
PyMuPDF>=1.25.4

# Supabase
supabase==1.0.3
postgrest-py==0.10.3

# Notifications
sendgrid>=6.11.0
twilio>=8.10.2

# Utilities
python-dotenv>=1.0.0
pydantic<2.0.0,>=1.9.0
tenacity>=8.2.3
python-magic>=0.4.27
aiofiles>=23.2.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.1
pytest-cov>=4.1.0
httpx>=0.23.0,<0.25.0 